backend/requirements.txt
click==8.0.0
Flask==2.0.0
Flask-Cors==3.0.10
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
marshmallow==3.12.1
Paste==3.5.0
psycopg2-binary==2.8.6
six==1.16.0
waitress==2.0.0
Werkzeug==2.0.0
